+++ src/SmoothNanners.Web/Views/Error/Index.cshtml
@@ -1,4 +1,4 @@
-@model SmoothNanners.Web.Models.Home.ErrorViewModel
+@model SmoothNanners.Web.Models.Error.ErrorViewModel
 @{
     ViewData.SetTitle($"Error: {Model.Code}");
 }
